pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bayesian_memory_ladder"
version = "0.2.0"
description = "Run-disciplined suite for non-Markovian primordial memory models with CAMB/CLASS/ModeCode."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "pandas>=2.0",
  "scipy>=1.10",
]

[tool.setuptools.packages.find]
where = ["src"]
